Apply new project structure
This commit changes the structure of project packages. A naming convention is followed in which package names are singular nouns with lowercase letters. The __init__.py files of packages will contain code. The new structure may make it simpler for Python newcomers to track simulator entities; at least, while the project is still lightweight.